se actualiza el float por un double

+++ TpFinal/src/busqueda/PorMontoMaximo.java
@@ -5,25 +5,24 @@ import prestamos.Prestamo;
 
 public class PorMontoMaximo extends Condicion{
 	
-	private float maximo;
+	private double maximo;
 
-	public float getMaximo() {
+	public double getMaximo() {
 		return maximo;
 	}
 
-	public void setMaximo(float maximo) {
+	public void setMaximo(double maximo) {
 		this.maximo = maximo;
 	}
 	
-	public PorMontoMaximo(float maximo){
-		
+	public PorMontoMaximo(double maximo){
 		this.setMaximo(maximo);
 		
 	}
 
 	@Override
 	public boolean respetaCondicion(Prestamo p){	
-			float montoAux = p.getMontoTotal();
+			double montoAux = p.getMontoTotal();
 			return (montoAux< this.getMaximo());
 	}
